+++ CHANGELOG.md
@@ -94,13 +94,21 @@ All notable changes to this project will be documented here. The format is based
 
 - There is now a 'translate to mesh bounds center' menu item in the mesh importer (#88)
 
-- There is now a 'translate to mesh average center' menu item in mesh importer:
+- There is now a 'translate to mesh average center' menu item in the mesh importer:
 
   - It takes the average center by adding up all the vertices and dividing by the number
     of vertices
 
   - This can sometimes be a handy approximation of "mass center", when the mesh isn't a
     closed surface (a requirement, if you want to compute an actual mass center)
+
+- (Experimental) There is now a 'translate to mesh mass center' menu item in the mesh importer:
+
+  - It uses "signed tetrahedron volumes" to do this, rather than something expensive like
+    voxelization
+
+  - This means that the resulting mass depends on the provided mesh being a closed surface
+    with correct normals etc.
 
 
 ## [0.2.0]